Moved urlopener to its own module
