+++ sc/app/objects.py
@@ -2,12 +2,6 @@
 This program holds all the non-database objects used necessary for
 ShakeCast to run. These objects are used in the functions.py program
 """
-try:
-    import urllib2
-except:
-    import urllib.request as urllib2
-
-import ssl
 import json
 import os
 import sys
@@ -21,46 +15,6 @@ from notifications import NotificationBuilder, Mailer
 modules_dir = os.path.join(sc_dir(), 'modules')
 if modules_dir not in sys.path:
     sys.path += [modules_dir]
-
-class URLOpener(object):
-    """
-    Either uses urllib2 standard opener to open a URL or returns an
-    opener that can run through a proxy
-    """
-    
-    @staticmethod
-    def open(url):
-        """
-        Args:
-            url (str): a string url that will be opened and read by urllib2
-            
-        Returns:
-            str: the string read from the webpage
-        """
-
-        # create context to avoid certificate errors
-        try:
-            ctx = ssl.create_default_context()
-            ctx.check_hostname = False
-            ctx.verify_mode = ssl.CERT_NONE
-        except:
-            ctx = None
-
-        try:
-            if ctx is not None:
-                url_obj = urllib2.urlopen(url, timeout=60, context=ctx)
-            else:
-                url_obj = urllib2.urlopen(url, timeout=60)
-                    
-            url_read = url_obj.read()
-            url_obj.close()
-
-        except Exception as e:
-            raise Exception('URLOpener Error({}: {}, url: {})'.format(type(e),
-                                                             e,
-                                                             url))
-
-        return url_read
 
   
 class AlchemyEncoder(json.JSONEncoder):
