+++ sc/app/productgrabber.py
@@ -3,7 +3,7 @@ import json
 import os
 import time
 
-from objects import URLOpener
+from urlopener import URLOpener
 from grid import ShakeMapGrid
 from orm import Event, Group, Product, ShakeMap, dbconnect
 from util import SC
